Suppress help for remote subparser
The remote subparser exists to re-parse only the necessary arguments for
issuing a remote command. This command should be handed off as-is to the
remote p instance, minus any command line arguments used to invoke the
remote execution. This means that --help, if present in the command,
should be ignored by the remote parser and instead passed to the remote
p instance.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>